requirements.txt
confluent-kafka==2.3.0
contextlib2==21.6.0
kafka==1.3.5
python-dotenv==1.0.0
PyYAML==6.0.1
schema==0.7.5
websocket-client==1.6.4
